wrap: Allow same syntax as machine files
For consistency it is better to quote values in wrap files and use
python style lists. It also allows using + and / operators as well as
[constants] section which is useful for example to write the project
version in a single place and use it in both source_url and
fallback_source_url.

For backward compatibility a legacy mode is added to
ConfigFileInterpreter that disable completely value parsing if it
fails to parse the first value (e.g. it is unquoted).