coredata: flatten compiler_options dict
This is a revert of https://github.com/mesonbuild/meson/pull/5489 that
was made for no clear purpose and benefices. This fixes Xc16CCompiler,
C2000CCompiler, C2000CPPCompiler, CudaCompiler and EmscriptenMixin that
were not ported.

This is in preparation to add per-subproject compiler options that will
also include a subproject prefix to option names.