Remove uneeded `parse_cmd_line_options`
The new way of doing things, thanks to @dcbaker, is making it so
`Environment` combines the command line options and config files, and
then coredata and just ingestion 1 source of raw data to initialize the
actual options. This dramatically simplifies things by making
information flow through one path not many.

`parse_cmd_line_options` is a vestige of the old way of coredata having
to crawl over a bunch of different sources of information and repeatedly
parse them, and is no longer needed.