avoid spurious warning from get_target_filename
In the situation covered by the new test case, get_target_filename is
invoked in the process of retrieving (via get_target_private_dir) the
custom target's @BUILD_DIR@; because the custom target has more than
one output, a spurious warning is printed.

However, that substitution is not used in the commands.  Avoid the
problem by not invoking get_target_private_dir.

This seems to be a regression from 0.54.3 to 0.55.0.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>