qt dependency: do not require all the tools automatically
The compilers_detect function is only used in the qt module, which
checks every time before using a specific compiler, if it is found.

e.g.

  meson.build:10:6: ERROR: MOC sources specified and couldn't find moc-qt5, please check your qt5 installation

In fact, the current check means we never even hit this error to begin
with, because we previously died on the uninformative error:

  meson.build:10:6: ERROR: Program(s) ['moc'] not found or not executable

which doesn't actually tell the user why this matters, and is all around
a waste of time.

Fixes #5582