cmake: Use a mapping when writing compiler ID
Meson and CMake compiler ids are different. This commit adds a mapping
from the meson list:

https://mesonbuild.com/Reference-tables.html#compiler-ids

to the CMake list:

https://cmake.org/cmake/help/latest/variable/CMAKE_LANG_COMPILER_ID.html

The mapping is not 1-1, and not all entries are mapped, so this is
a best-effort attempt. Fallback to GNU as before to try to limp along
and hope that the build files don't rely on an accurate compiler ID.