Rewrite buildtype as debug+optimization pair in user inputs
This deprecate get_option('buildtype') to avoid the trap when it's set
to 'custom'. It is still useful for the user to write
-Dbuildtype=release as a way to declare intention rather than having to
dive into optimization level details.

Note that this makes mixing buildtype and debug/optimization in command
line or in default_options a hard error, but not when using buildtype in
command line and debug/optimization in default_options (or vice versa).
In other words, `default_options: ['buildtype=debugoptimized']` with
`-Ddebug=false` is not an error and results in buildtype='custom',
debug=false and optimization=2.