Fix picking up tools with args from the env on Windows
This was a regression in https://github.com/mesonbuild/meson/pull/7059.
We do not need to do a full search for all windows special cases. We
only want to check whether the command provided is actually a full
path to a script so we can add the interpreter.

Fixes https://github.com/mesonbuild/meson/issues/7645