Revert "msvc: enable /std:c11 flag"
This reverts commit b5edb514d7e7a4a6852167bcccf57940ab245ccf, which
was not supposed to be cherry-picked to the stable branch.

It causes projects that were previously working and had c_std=gnu99 in
default_options to error out with:

```
meson.build:1:0: ERROR: Value "gnu99" for combo option "C language
standard to use" is not one of the choices. Possible choices are:
"none", "c89", "c99", "c11".
```