Add find_program and override_program to environment
This provides a single entrypoint for finding all programs, external and
internal, in meson. It provides caching of found results (but not of
not-found results), as well as a single method to store and retrieve
overridden programs. It uses two separate caches, a persistant cache in
coredata, which is used for external programs, and a transient cache
used for overridden programs that lives in the environment class.

I konw that placing this in Environment will be somewhat controversial,
as that class is already very large and complicated. However, I think
it's the right place for two reasons:

 1. The persistant cache belongs in the coredata, it needs to be
    serialized and retrieved to be useful.

 2. The transient cache belongs in Environment. Environment is supposed
    to be for information that needs to be recomputed each time that
    meson is run. overriden programs need this, as they could change
    between invokations.

Given that find_program already needs environment and coredata
references to function, it makes sense to make it part of Environment,
and not pass those in to each call.