stabilize iteration order for dictionaries
The order of keys in dictionaries cannot be relied upon, because the hash
values are randomized by Python.  Whenever we iterate on dictionaries and
meson.build generates a list during the iteration, the different iteration
orders may cause random changes in the command line and cause ninja to
rebuild a lot of files unnecessarily.