[WIP] Add find_program(..., fallback: subproject) support
This patch adds support for explicit subproject fallback support for
executables searched with `find_program`. This allows vendored
dependencies to be shipped in `subprojects/` but prefer system packages
when available.

The use of this flag is also independent of whether wrap downloading is
configured allowing simple explicit vendoring to be more flexible

TODO: Update the manual