mtest: store TestRuns in collected_failures
Instead of storing the string, store the whole TestRun.  In the
next patches we'll use this to colorize the summary of failures,
and to allow a few more simplifications.

There is some code duplication between the console and logfile
code, but it won't matter as soon as console and logfile output
will be in two completely separate classes.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>