mtest: introduce TestLogger
The TestLogger class lets us move the code for all those log files
out of TestHarness.  The interface is based on JunitBuilder, which
is converted already in this commit.  Over the next commits, we
will also convert JSON, text and console output.

The main difference with JunitBuilder is that the completion method is
asynchronous.  This can be useful if the logger needs to clean up after
itself and wait for asyncio tasks.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>