mtest: remove collected_logs
Just reuse the collected_failures collection now that it contains
TestRun objects.  Move the code to generate the short form of the log
to TestRun.

Note that the first line of the error log is not included in
get_log()'s return value, so the magic "first four lines are passed
unscathed" is changed to three lines only.  The resulting output is
like this:

    --- command ---
    <command line>
    --- Listing only the last 100 lines from a long log. ---
    --- stdout ---
    ...

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>