mtest: use backslash replace when printing error logs
If there's an UnicodeEncodeError while printing the error logs,
TestHarness tries an encode/decode pair to get rid of anything that
is not a 7-bit ASCII character; this however results in "?" characters
that are not very clear.  To make it easier to understand what is
going on, use backslashreplace instead.

While at it, fix the decode to use a matching encoding.  This will
only matter in the rare case of sys.stdout.encoding not being an
ASCII superset, but that should not matter.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>