doc: fix confusing use of "build" as a directory argument [skip ci]
In most places, we now refer to "builddir/" which is a lot less likely
to make people think it is a subcommand which needs to be used
literally.

This is a regression since commit 276d342ebaf859dd53e145ead3e98e2cebb360ab
due to the existence of new docs which were added later on, using the
wrong form.