external-project: Add executable() method
Add an executable() method constructing an object pointing to a
program built as part of the external project build and suitable to be
passed to meson.override_find_program(). This allows to use external
projects to build code generators to be used in other parts of the
build.

Note that the program is run from the temporary install location in
the build directory, thus it could fail to run if it depends on
resources located at specific absolute filesystem paths. In
particular, binaries relying on shared libraries built as part of the
external project would fail to load. This could maybe be solved using
an appropriate LD_LIBRARY_PATH environmental variable, but no effort
in this direction is done here.