Add fortran test with an install:yes static library
This adds a test which makes use of an install:yes static library that
depends on another static library. This triggers a promotion to
link_whole_target inside meson which takes different code paths in
certain places.

Also makes use of .F90 source (capital F) to test for case
(in)sensitivity.