Better documentation and argument checking for multiple install_dirs
The fact that each output can be sent to a different directory isn't
documented, which is bad. The fact that `false` can be used to disable
installing specific outputs isn't documented either, which is bad. The
code doesn't do any sort of checking that the values are valid, so True
will be accepted, which shouldn't be (and probably leads to a bug
later).

On top of that, the checking for this is done in the ninja backend, not
in the interpreter. That's a huge layering violation, and means that if
you don't use ninja invalid code will be allowed through.

Lets fix all of that.