custom_target: Add a '@MANDIR@' special option for install_dir
This simplifies installing generated man pages from custom_targets.
Instead of manually passing the man page information, you can now set:
`install_dir : '@MANDIR@'`, and Meson will do this for you.

This is a sort of compromise between passing generated data to
`install_man` and having to fully calculate it by hand. It keeps with
Meson's design to provide all information to a target when that target
is defined, but not forcing users to calculate the names of man files.
It does have one advantage, in that it can more elegantly handle a
custom_target that generates both man pages and other kinds of docs at
the same time:

```meson
custom_target(
 ...,
 output : ['man.1', 'man.html'],
 install_dir : ['@MANDIR@', docdir],
)
```

Fixes: #1550